JANUS-257: Started work on tile beam metadata

+++ src/ska_low_mccs_daq/cdaq/BeamformedData.cpp
@@ -102,9 +102,9 @@ bool BeamformedData::processPacket()
     uint64_t payload_length = 0;
     uint64_t sync_time = 0;
     uint64_t timestamp = 0;
-    uint16_t beam_id = 0;
+    uint8_t beam_id = 0;
     uint16_t frequency_id = 0;
-    uint16_t tile_id = 0;
+    uint8_t tile_id = 0;
     uint16_t station_id = 0;
     uint16_t nof_contributing_antennas = 0;
     uint32_t payload_offset = 0;
@@ -152,7 +152,7 @@ bool BeamformedData::processPacket()
             case 0x2003: // Tile and Station information (LMC data)
             {
                 uint64_t val = SPEAD_ITEM_ADDR(item);
-                tile_id    = (uint16_t) ((val >> 32) & 0xFF);
+                tile_id    = (uint8_t) ((val >> 32) & 0xFF);
                 station_id = (uint16_t) ((val >> 16) & 0xFFFF);
                 nof_contributing_antennas = (uint16_t) (val & 0xFFFF);
                 break;
@@ -280,9 +280,9 @@ bool IntegratedBeamformedData::processPacket()
     uint64_t payload_length = 0;
     uint64_t sync_time = 0;
     uint64_t timestamp = 0;
-    uint16_t beam_id = 0;
+    uint8_t beam_id = 0;
     uint16_t frequency_id = 0;
-    uint16_t tile_id = 0;
+    uint8_t tile_id = 0;
     uint16_t station_id = 0;
     uint16_t nof_contributing_antennas = 0;
     uint32_t payload_offset = 0;
@@ -330,7 +330,7 @@ bool IntegratedBeamformedData::processPacket()
             case 0x2003: // Tile and Station information (LMC data)
             {
                 uint64_t val = SPEAD_ITEM_ADDR(item);
-                tile_id    = (uint16_t) ((val >> 32) & 0xFF);
+                tile_id    = (uint8_t) ((val >> 32) & 0xFF);
                 station_id = (uint16_t) ((val >> 16) & 0xFFFF);
                 nof_contributing_antennas = (uint16_t) (val & 0xFFFF);
                 break;
